Cs11/backend/requirements.txt
asgiref==3.5.2
Django==4.0.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
Pillow==9.4.0
gunicorn==20.1.0
django-filter==23.2
